requirements.txt
psycopg2_binary==2.9.6
loguru==0.7.0
pika==1.3.2